Added confirmation popup for dangerous actions

+++ src/gui/content.rs
@@ -98,6 +98,7 @@ pub fn nothing_selected(app: &mut FmpApp, ui: &mut egui::Ui) {
     });
 }
 
+#[allow(clippy::too_many_lines)]
 /// Displays the content for the main window of the application when a vault is selected.
 ///
 /// # Arguments
@@ -166,7 +167,6 @@ pub fn vault_selected(app: &mut FmpApp, ui: &mut egui::Ui) {
         });
 
         ui.separator();
-
         ui.add_space(8.0);
 
         ui.horizontal(|ui| {
@@ -178,20 +178,26 @@ pub fn vault_selected(app: &mut FmpApp, ui: &mut egui::Ui) {
                 .button(egui::RichText::new("Delete Vault").color(egui::Color32::LIGHT_RED))
                 .clicked()
             {
-                match delete_vault(app) {
-                    Ok(()) => {
-                        app.toasts
-                            .success(format!("Vault `{}` deleted.", app.vault_name))
-                            .duration(Some(Duration::from_secs(2)));
+                app.show_confirm_action_popup = true;
 
-                        app.vault_name.clear();
-                        app.fetch_vault_names();
+                if app.confirm_action {
+                    match delete_vault(app) {
+                        Ok(()) => {
+                            app.toasts
+                                .success(format!("Vault `{}` deleted.", app.vault_name))
+                                .duration(Some(Duration::from_secs(2)));
+
+                            app.vault_name.clear();
+                            app.fetch_vault_names();
+                        }
+                        Err(_e) => {
+                            app.toasts
+                                .error(format!("Failed to delete vault `{}`", app.vault_name))
+                                .duration(Some(Duration::from_secs(3)));
+                        }
                     }
-                    Err(_e) => {
-                        app.toasts
-                            .error(format!("Failed to delete vault `{}`", app.vault_name))
-                            .duration(Some(Duration::from_secs(3)));
-                    }
+
+                    app.confirm_action = false;
                 }
             }
         });
@@ -205,7 +211,6 @@ pub fn vault_selected(app: &mut FmpApp, ui: &mut egui::Ui) {
         });
 
         ui.separator();
-
         ui.add_space(8.0);
 
         backup(app, ui);
@@ -303,27 +308,31 @@ pub fn account_selected(app: &mut FmpApp, ui: &mut egui::Ui) {
             .button(egui::RichText::new("Delete Account").color(egui::Color32::LIGHT_RED))
             .clicked()
         {
-            match delete_account_from_vault(app) {
-                Ok(_o) => {
-                    app.toasts
-                        .success(format!(
-                            "Account `{}` deleted from vault `{}`.",
-                            app.account_name, app.vault_name
-                        ))
-                        .duration(Some(Duration::from_secs(2)));
+            app.show_confirm_action_popup = true;
 
-                    app.account_name.clear();
-                    app.clear_account_data();
-                    app.fetch_account_names();
-                }
+            if app.confirm_action {
+                match delete_account_from_vault(app) {
+                    Ok(_o) => {
+                        app.toasts
+                            .success(format!(
+                                "Account `{}` deleted from vault `{}`.",
+                                app.account_name, app.vault_name
+                            ))
+                            .duration(Some(Duration::from_secs(2)));
 
-                Err(e) => {
-                    app.toasts
-                        .error(format!(
-                            "Failed to delete account `{}` from vault {}. Error: {}",
-                            app.account_name, app.vault_name, e
-                        ))
-                        .duration(Some(Duration::from_secs(3)));
+                        app.account_name.clear();
+                        app.clear_account_data();
+                        app.fetch_account_names();
+                    }
+
+                    Err(e) => {
+                        app.toasts
+                            .error(format!(
+                                "Failed to delete account `{}` from vault {}. Error: {}",
+                                app.account_name, app.vault_name, e
+                            ))
+                            .duration(Some(Duration::from_secs(3)));
+                    }
                 }
             }
         }
@@ -502,6 +511,11 @@ pub fn alter_vault_name(app: &mut FmpApp, ui: &mut egui::Ui) {
     });
 }
 
+/// Displays the content for generating a password.
+///
+/// # Arguments
+/// * `app` - A mutable reference to the `FmpApp` instance containing the application state.
+/// * `ui` - A mutable reference to the `egui::Ui` instance for rendering the user interface.
 pub fn random_password(app: &mut FmpApp, ui: &mut egui::Ui) {
     ui.add_space(12.0);
